Added debug port and improved exception handling in Dockerfile
The changes in the Dockerfile focus on enhancing the installation process of pydevd-odoo and its setup within the system. An environment variable ODOO_DEBUG_PORT was introduced to provide more flexibility for debugging. The inclusion of a try/exception block is significant to avoid disruptions that may occur during the setup phase. It captures and prints exceptions that could prevent the successful execution of the import command and settrace function. These changes lead to a more stable setup, making errors easier to trace and rectify.